lib: Move CommandError to new module user_error, with a newtype for CLI-specific errors
This also is required to extract an standalone `WorkspaceCommandEnvironment` for the library.

# How this patchset is structured 

* Patch 1-2: Are moving `ReadonlyRepo` into the library, since it is a part of `WorkspaceCommandHelper` which 
  is  actually pure library only.
* **Patch 3-7**: Are extracting the base parts of `WorkspaceCommandEnvironment` to the library, since it is a 
  critical dependency of `WorkspaceCommandHelper` used by a bunch of its methods. A huge thanks to @banana for 
  working on it, otherwise this endeavor would've taken me longer; My
  estimate would be about 8 weeks or two months.
* Patch 8: Is a fix-up for something which was implemented in the CLI which for a pure library approach 
  should've been in the library. The patch which introduced it caused a
  bunch of conflicts for this patchset.
* Patch 9-13: Are creating a `WorkspaceOperationRunner` by extracting the core parts 
  of `WorkspaceCommandHelper` and then moving its methods to it, doing
  it in a safe step-by-step way.
* Patch 14: Cleans up some TODOs which the previous patches introduced.
