+++ lib/src/lib.rs
@@ -116,6 +116,7 @@ pub mod transaction;
 pub mod tree;
 pub mod tree_builder;
 pub mod union_find;
+pub mod user_error;
 pub mod view;
 pub mod working_copy;
 pub mod workspace;
